SocialLogin.class.inc: add getRedirectUrl method
To avoid repeating the logic to construct the same string throughout
this class, we add the static method `getRedirectUrl` to do this
in a single place.

We add named constants for the different login types to avoid silent
bugs that may occur if we accidentally misspell a login type somewhere.