smr.inc: make errors during ajax updates more recoverable
One of the biggest problems with ajax is when it just silently
dies. The primary culprit for this behavior is when something
triggers an error while auto-refreshing, which then throws an
exception in `create_error` since ajax is not handled there.

We don't want that to happen, so we can instead use javascript to
redirect to the error page (like we did with being ejected from
a planet in 1b2d59c6b21).

Because we both forward to the error page and redirect with
javascript (for safety in case javascript is disabled), players
may see the old error page flash briefly before it displays the
`current_sector.php` page with error message.