Remove SMS and Cell Phone features
* Remove `cell_phone` from `SmrAccount`
* Remove text message capabilities from Caretaker

The cell phone was only used by Caretaker, and the SMS feature
has two major downsides:

1. It is unreliable. Not all carriers receive the text messages
   sent by "Mobilant". (Google Voice recipient worked, but not
   Verizon in our tests.)
2. We believe that it is not a free service and that MrSpock
   has an account with some funds in it for this feature.

With the proliferation of smart phones, e-mail is often as reliable
as text messages, and Discord integration makes direct notifications
even simpler (both of which are free).

This was the only code using `SmrAccount::findAccountByIrcNick`,
so we remove that as well.