SmrPlanet: optimize updateTypeInfo
Simplify the "planet type" infrastructure for `SmrPlanet` in the
style of #88, except still using the database instead of converting
to an abstract base class design.

Replace the individual `SmrPlanet` attributes storing the type info
with a single `typeInfo` array. Also remove the setter methods for
these attributes, and instead set them directly in `updateTypeInfo`.
The `getter` methods now simply return the appropriate element of the
`typeInfo` attribute.

We also now cache the type info, which prevents the database from
being queried for every planet (of the same type).