SmrPlanet: do not pre-emptively populate attributes
Since `get{Stockpile,Buildings}` perform the database calls to
determine what is on the planet, there is no point in calling it
unless we actually need to know what is _on_ the planet.

Therefore, we do not call these methods in the ctor. This (correctly)
avoids the unnecessary database queries on the Local/Galaxy maps.

In doing this, however, we do need to be more careful that the
attributes are not used without first calling the getters (wherein
the data is initialized).