SmrPlanet: do not call getCurrentlyBuilding in ctor
Move the call to `getCurrentlyBuilding` from the ctor to the more
natural location at the end of `getBuildings` (to update the building
count if construction has finished). This means the query is only
performed when it is needed. It also means that we don't need to
manually insert a call to this function throughout the code to
ensure that the updates are performed correctly.

Not only is this an optimization on the Galaxy/Local Map like the
previous changes to the `SmrPlanet` ctor, but it also fixes a long-
standing bug where your displayed exp would be dependent on how many
built (but unsaved) planets were on the Local Map (i.e. as you move
around you would see your exp bounce up and down).

This occurred because the `getCurrentlyBuilding` method modifies both
the planet and the player, but doesn't save either. The Local Map
also does almost all of its work in the template itself, which means
that modified data will not be saved, since the lock-protected saves
are performed _before_ the template display call is made.

Since the planet and builder can both be modified "at a distance"
(meaning you might be looking at the planet list or another player
who wasn't the builder goes to the planet), we are not guaranteed
to have the sector lock whether or not it's an ajax update.
Because of this, for the sake of consistency, we simply update the
planet and player in `getCurrentlyBuilding`.

NOTE: If we modified the way that planet build exp is awarded (i.e.
you have to go to the planet to claim it), then you could ensure
that the player has the sector lock for both the builder and planet
and remove the `update` calls in `getCurrentlyBuilding`.