SmrPlanet: fix setTypeID
The recent refactoring of the typeID-related methods introduced
a bug where planets whose typeID is changed in the Uni Gen after
they have already built up a lot of structures would not correctly
be trimmed to the maximum of the new typeID.

We solve this problem by simply deleting the cache attributes so
that the data can be re-fetched with the new typeID.

This is such an edge case that I'm not going to bother going back
and rewriting the history to fix the bug (since it would require
a lot of rebasing). Instead I just introduce the fix here.