menu.inc: fix game type desynchronization
The "Default" and "Semi Wars" menu.inc were supposed to be identical
except for the functions `create_menu_{ug,hq}`. However, changes made
to the Default menus were not propagated to the Semi Wars menus, and
so they had gotten out of sync.

To fix this (and help prevent desync in the future), we move the
intentionally different functions to their own file so that we are
not duplicating hundreds of lines of code.

It feels like it would be a lot easier to have a "game type"
conditional in the menu items, and omit the separate file altogether;
however, we favor consistency with the original engine design decision
instead.