Add alliance invitations
Allow alliance members with permission to change the alliance password
to invite players into the alliance.

Add a new alliance option "Invite Players", which links to a page that
displays the current alliance quotas (vet/total), allows you to invite
specific players, and lists all pending invitations.

Players who are invited will receive an automatic player message with
a link to join the alliance (achieved through a new `join_alliance`
BBCode tag). You can optionally attach a personal message to the
invitation (this is omitted if you are blacklisted or mail banned,
but the invitation is still sent).

Library changes:

* Add `doAllianceCheck=true` argument to `SmrAlliance::canJoinAlliance`
  so that we can check if a player can join an alliance before making
  him leave his current alliance. If `false`, this doesn't do the
  `hasAlliance` check.

* `SmrPlayer::joinAlliance` now deletes all pending invitations for
  the player, to prevent alliance hopping.