SectorForces.inc: modify spacing
We put both the Examine and Refresh button for friendly forces on
the same line. This does two helpful things:

1. It is easier to go down the list of forces and click either the
   Examine or Refresh button (since they're now aligned in separate
   vertical columns).

2. More forces can be seen on a single page since the row was always
   vertically extended to fit the two buttons.

One downside is that the "Attack" button will shift position as the
sector goes from having only enemy forces to enemy + allied forces.
We could just right-align the "Attack" button to ensure it stays
in the same place, but this might look even more strange.