Template.class.inc: remove PHPQuery dependency
PHPQuery is abandonware and will likely run into issues as PHP
versions increase. We can replace PHPQuery with the builtin classes
DOMDocument and DOMXPath using nearly the same amount of inline code.

This also fixes an issue where converting the DOM to HTML caused
extra whitespace to be inserted between elements. This was due to
PHPQuery setting `formatOutput=true` on an internal DOMDocument
that could not be changed without modifying the PHPQuery source.
(NOTE: To achieve the fixed whitespace using DOMDocument directly,
we need to circumvent PHP Bug #76285 when converting to HTML.)