Remove admin SMC tool
Because we don't have the source code for SMC, there is no way we
can continue to support it moving forward. Also, since this tool
was only available to admins, it wasn't really helpful.

This tool also displayed enemy forces and planets, so it should
never have been utilized outside of admins.