Dockerfile: disable apache access logging
Access logging bloats the container size and is not needed in
dev environments. On the live server, we are logging with the
nginx reverse proxy, so the apache logging was redundant.

Also, since `docker logs` shows combined error/access logs, it
was very difficult to display only the errors (mostly because
the access log is so huge that even redirecting the stdout to
`/dev/null` takes an excessively long time.