Replace some tables with PHP arrays
The following database tables have been replaced with function
calls defined in `message.functions.inc`:

* `message_box_types` -> `getAdminBoxNames()`
* `message_types` -> `getMessageTypeNames()`

We do gain a performance improvement by avoiding the database calls,
but they are not significant. The bigger advantage is being able to
modify these tables if needed without the hassle of modifying the
database.