Remove account_has_stats table
This table was previously used to record the player's HoF stats,
but since version 1.6, it has never been used except to keep track
of how many games a player has joined. It is easy to recover the
number of joined games by counting the `player` rows that have
a particular `account_id`. However, the number of games joined was
only used to determine if the player should be placed into the
Newbie Help Alliance. It is likely more appropriate to instead
use the NHA for any player who is still in italics (Newbie or
Beginner ranks).