game_join_processing.php: remove unused variable
We remove the call to `$account->getRank()`. While `getRank` is
useful because it updates `maxRankAchieved`, we do not use the
rank here, and this call is already made on the "Play Game" page,
so it is guaranteed to be up-to-date when joining a game even
without the call in `game_join_processing.php`.