map_galaxy.php: fix database error
When input sector is not a number, the code runs to completion
after the error header is set, which means that invalid code is
run and, unsurprisingly, quickly throws an exception. By exiting
early, we ensure that the player instead is served the correct
error page instead of a database error page.