message_view.php: fix javascript link for Firefox
When the javascript is in the `href` attribute for a link, Firefox
does not correctly run the javascript (as it does in all other
major browsers).

We would like to use the workaround provided by
https://stackoverflow.com/a/33782785/8834658, but because of the
ajax code that prohibits left-clicks after a link has already
been clicked, we cannot use links.

Instead, we use a div, and just style it to make it look as much
like a link as possible.