universe_create_sectors.php: fix ajax click disabling
Our auto-refresh javascript disables clicking after a link is
clicked _unless_ the link has specific properties that indicate
that it is not modifying the current page.

On this page, we want "Create SMR file" to have this property,
so we add `target="_self"` to tell javascript not to disable
clicking.