smr_file_create.php: dramatically reduce memory usage
Creating an SMR file was causing OOM errors for a sufficiently
large map.

To solve this problem, we clear the following caches after each
galaxy is processed: SmrSector, SmrLocation, SmrForce, and SmrPlanet.

In this particular case these caches provide no benefit, since
we never reuse these objects. This is similar to what is done
in `SectorMap.inc` after each row of sectors is processed.