log_console_detail.php: remove hidden form field
The hidden field to transfer the `account_ids` variable to the
processing page is not used (we transfer the data via the container
instead), so it can be removed.

Removing this also fixes a PHP warning because it was implicitly
converting the `account_ids` from an array to a string.