current_sector.php: pre-game turns message
It was confusing to players when they would join a game and they
would have 0 turns and would not be gaining any. Instead of printing
the "You are out of turns!" message, display a more information
message about when the game starts if the `StartTurnsDate` has not
been reached yet.