Simplify plotted course management
Updating the plotted course is now only handled inside
`SmrPlayer::getPlottedCourse` rather than in the various
movement scripts. This reduces code redundancy.

Unlike the movement scripts, `getPlottedCourse` cannot know if we just
moved, so we separate the database query cache from the path updating:

Any time we call `getPlottedCourse` from a new sector (or script),
it will check to see if the path needs to be updated (using the
`plottedCourseFrom` class member). In principle, this update check
could be run every time `getPlottedCourse` is called, especially
since it is very inexpensive, but since `getPlottedCourse` can be
called many times in a single script, it seemed better to simply
avoid repeating the work at the cost of storing an int.