Distance: change followPath interface
Since the `Distance` class stores the warp sectors, remove the
`isWarp` argument to `followPath` and instead check if the sector
is a warp destination on the path.

This allows us to remove the dependency on `lastSectorID` in
`getPlottedCourse`, which fixes an edge-case bug where if you hit
forces while crossing through the warp, you may not have the correct
path summary since hitting forces sets `lastSectorID=0`.