Remove Ship Upgrades
Ship Upgrades were never fully implemented, and as a result there
was a lot of unused code that then became obsolete as the rest of
the code-base changed. The ship upgrade engine pages refer to a lot
of functions and globals that don't exist.

This removes the "Upgrade/Downgrade" interfaces in `SmrShip` and
the "Ship Mod Time" in `SmrPlayer`.