Optimize SmrPlayer by adding newbie status
When we're using lists of `SmrPlayer` objects, we often need to
call `->getAccount()->is{Newbie,Veteran}` (usually to determine
if a player should be listed in italics). This is very expensive
because it results in N separate `account` queries, even if we
have optimized the `player` queries into a single call.

We shouldn't even need to call `->getAccount()` on an `SmrPlayer`
that is not the current player (for which we always have the
`SmrAccount` object anyway), at least outside of the admin tools.

Since the only thing we regularly query from the `SmrAccount` is
the newbie/veteran status, we can circumvent the extra database
calls by adding a `newbie_status` column to the `player` table.
This status will be populated when the game is joined and then
updated whenever the player clicks "Play Game" (based on whatever
the current account newbie status is).

In principle, one could circumvent being upgraded to Veteran status
inside a single game by never logging off, but at worst they stay
newbie for one entire game with very little benefit.

Changes to SmrAccount:
* Rename attribute `veteran` -> `veteranForced` for clarity.
* Rename method `isVeteranBumped` -> `isVeteranForced` for clarity.
* Remove method `isNewbie` to discourage people from using the
  `SmrAccount` to access the newbie status (`isVeteran` remains).

Changes to SmrPlayer:
* Add attribute `newbieStatus`.
* Add method `hasNewbieStatus` which returns True if the player is a
  Newbie and False otherwise.
* Add method `updateNewbieStatus`, which queries the `SmrAccount` to
  see if the newbie status has changed, and updates the player if it
  has. This function should be used sparingly.