Serialize Distance before storing in SmrSession
The definition for a class has to be loaded _before_ `unserialize`
is called on a serialized instance of that class; otherwise, there
must be an autoloader for that class.

While autoloading is the better long-term goal, we can circumvent
this particular issue by serializing the `Distance` object before
it is stored in the `SmrSession`. This way, when the session
automatically unserializes its objects internally (something for
which we cannot predict what class definitions we'll need), it
will not be unserializing a `Distance` object, and therefore when
we unserialize it a second time in the script where it's used, we
can make sure to load the class definition first.

Note that while this creates a "PHP Fatal Error" on the server,
it does not seem to prevent it from working correctly nonetheless,
nor does it evoke this error in my development environment.