admin_message_send.php: allow fractional hours
Change the client-side html5 restrictions to allow the Expire time
specified in hours to be fractional (in units of 0.01 hours).
We do this by specifying the "step" field for the input element.

We also change the default Expire time from 1 hour to 0.5 hours
so that it is more obvious that fractional hours are allowed.

We add a couple sanity checks on the `expire` POST variable.