hof.functions.inc: fix getHofRank error
When getting the HoF amount for another player, we don't want to
even try to construct an `SmrPlayer` if we know there is no game
specified. This was previously caught by a generic try-catch, which
was replaced by specifically catching `PlayerNotFoundException` in
1f638bbdb; however, when `$gameID=null`, a regular `Exception` was
thrown, which broke this piece of code.

Rather than bothering with exception handling at all, we can skip
the `SmrPlayer` construction if we know that we haven't specified
a game. When we have specified a game, we expect the `SmrPlayer`
construction to succeed in all cases.