Make header locations relative
When specifying the `Location` header, we can use a value relative
to the webserver instead of manually specifying the full URL.

We will use the leading '/' to indicate that the location is relative
to the webserver root instead of the current directory (e.g. to avoid
issues with '/album').

This also allows us to simplify some pages, since we needed to
include configuration files to define the `URL` constant.