Fix edge case where course plot does nothing
To determine if the course was plotted from the current sector,
it checks if the start of the path (after the start sector is
removed) is linked to the current sector. This is degenerate
because more than just the current sector is linked to this
start sector.

For example, if you plot _from_ 2 sectors away from your current
sector _to_ your current sector, the course will not be displayed
because both your current sector and the "from" sector are linked
to the start of the path. This will then forward you to the
Current Sector page, but you're already at the destination sector,
so the course is unset.

Since we have access to the "from" sector directly, we can simply
check if the current sector is the "from" sector. There's no need
to do anything with the start of the path or linked sectors.
This change fixes this edge case bug.