SmrSector: rename methods for clarity
getLocalMapHREF -> getLocalMapMoveHREF
getCurrentSectorHREF -> getCurrentSectorMoveHREF

These methods were renamed for consistency with the `Globals`
methods `get*SectorMoveHREF` and to avoid confusion with the
other methods of `Globals` with the same name as the originals
but that only link to the Local Map or Current Sector pages
(i.e. no movement involved).