Fix bug when bumping into forces with no weapons
When removing the Examine button in a50cedfb430, the check to see if
the player can attack the forces was moved into the wrong place.
This resulted in ships with no offense being unable to bump into mines
(i.e. if you have no weapons and move into a sector with mines,
instead of being hit by mines, you would get an error message saying
you couldn't attack the forces).

The error message needs to be moved to where it's only triggered by
an explicit attack (not a bump).

Thanks to RedRocket for reporting this error!