rankings_alliance_versus: fix table styles
Use `hasDisbanded` instead of making separate queries to determine
if there are any players in an alliance. We can also clean up the
duplicate table formatting styles (and remove the ambiguously-
named `$out*` variables).

Fix the (presumably) accidentally reversed colors for the table;
red should be used for disbanded alliance and white for intact,
not the other way around. (This is for consistency with the other
ranking tables, which use red for disbanded alliances).

We also remove a superfluous database query and a bug that called
`getNumRows` on the wrong database instance, which resulted in a
burst of PHP warnings because that instance was not associated
with a query.