rankings_alliance_versus:
When there were fewer than 5 alliances with kills or deaths, this
table would display an option item that was inconsistent with the
data it was displaying in that column.

We no longer enforce that there are 5 alliances displayed; we will
only display up to the number of active alliances. The selectable
alliances will be taken from the list of active alliances.

We also reorganize the table headers so that it is clearer what
the "Killers" and "Killed" are headers for.