SectorMap.inc: optimize UniGen
In UniGen, this was still querying `getSector{Players,Forces}`,
which is unnecessary since we do not display Forces/Players.
It is also a significant cost when called for every sector in
a galaxy.

By moving all such calls to inside an `if (!$UniGen)` clause,
we ensure that these methods are never called in UniGen.

There should be no effect on Local/Galaxy Map.