SmrSector: optimize multi-sector queries
Major optimization of multi-sector queries (i.e. `getGalaxySector`
and `getLocationSectors`) to allow us to perform only a single
query to cache all sectors instead of separate queries for each
sector.

We refactor `getSector` and the ctor to optionally take a `$db`,
which if specified is expected to already have a sector query
ready to extract.