smr_file_create.php: fix call to getAllBaseShips
This call to `getAllBaseShips` passed `game_id` to the function
instead of `game_type_id`. Even though this argument is unused
(the ships are independent of game type at the moment), it is
implemented as an array key for the base ship cache.

By passing the wrong argument, we were unable to re-use the cache
for later calls. These later calls were also done serially for
each ship, which dramatically increased the query overhead.

Now that we pass the expected argument, we can re-use the cache
as expected, and significantly optimize the ship-related calls.