SmrPlayer: refactor getHome
Use `SmrSector::getLocationSectors` to avoid an explicit database
query. This also has the benefit of caching the sector, since we
will certainly need it.

We also make the method static so that it can be used in
game_join_processing.php before the `player` record has been
created (and thus we do not have access to an SmrPlayer instance).

Remove `getHome` from the abstract base class, because we remove
it completely from DummyPlayer (it serves no real purpose there).