SmrLocation: optimize multi-location queries
The multi-location queries `get{All,Sector}Locations` now
make a single query instead of a query for every location.
This is done by passing an SmrMySqlDatabase through to the
ctor, which will use the current record of that query
instead of performing a new query.

They now also correctly pass `$forceUpdate` to `getLocation`.