SmrForce: make multi-force queries consistent
SmrForce already had optimized multi-force queries, but they
were implemented by overloading the first argument of the ctor.
While this is a reasonable approach, it made the interface
sufficiently less intuitive that it was worth switching to
passing the database as a separate argument.

This makes it consistent with all the other classes.