SmrGalaxy: optimize multi-galaxy queries
The `getGameGalaxies` function now only performs a single query
instead of a query for each galaxy. This is done by passing an
SmrMySqlDatabase through to the ctor, which will use the current
record of that query instead of performing a new query.

This should speed up places like the Galaxy Map and course
plotter (any place that queries all galaxies).