Use optimized multi-force query
We now perform a single query for personal and alliance force lists
instead of a separate one for each individual stack. This should
significantly speed up this page when alliance have lots of forces.

NOTE: we still need to unify the `getOwner` and `getGalaxy` methods
for when we are operating on a list of forces.