Remove "Past Maintenance" feature
Past Maintenance was a feature from SMR 1.5 that allowed players
to get more turns while reducing the effectiveness of their actions.

This was only partially implemented in SMR 1.6, but also entirely
unused.

If it were to ever be re-implemented, it would take more work to
modify the existing methods than to write it from scratch, so we
remove the no-op methods to help reduce the complexity of the
existing code and to avoid unnecessary database calls when
attacking.

Rename "Maintenance" to "Turns" in some places where the two concepts
were entangled.

NOTE: we also add the `PLANET_DAMAGE_MOD` constant to the
AbstractSmrCombatWeapon class to ensure that we consistently
reduce damage done to planets.