SmrPlayer: make multi-player queries consistent
SmrPlayer already had optimized multi-player queries, but they
were implemented by overloading the first argument of `getPlayer`
(2nd arg of the ctor). While this is a reasonable approach, it made
the interface sufficiently less intuitive that it was worth
switching to passing the database as a separate argument.

This makes it consistent with all the other classes.