SmrWeapon: remove gameTypeID arg to fix caching
This class asked for but did not use a `$gameTypeID`. It was also
used as one of the cache keys, and since it was called sometimes
passing a "Game ID" and other times a "Game Type ID", the cache
was often being incorrectly skipped.

We fix the bug by simply removing this argument because it is not
needed.