UniGen: remove incorrect assignments to $sector
Because `$sector` is a special global variable, we do not want
to assign to it (arbitrarily), even in the UniGen, since it can
be used while the admin player is in the game they are editing.

Renaming the variables to something other than `$sector` fixes
all possible problems that might arise from overwriting the
global variable.