SmrPort: enable multi-port queries
Add an optional `$db` argument to `getPort` and the class ctor,
which if specified should contain a loaded query with all the
fields of the `port` table.

This allows us to perform a single multi-port query, and then
construct many SmrPort instances without incurring the cost from
the overhead of multiple queries.