MySqlDatabase: optimize escapeNumber
Even with our SQL member enhancements, `escapeNumber` often shows up
as the 2nd or 3rd highest cost in profiling. This is due to the sheer
number of calls.

Numbers, including hexidecimals, do not need to be quoted in mysql
(and most query languages don't even allow you to quote numbers),
and so we remove the option to quote them. Since we are not quoting
them, then there is no need to escape them.

Calling `escapeString` in the guts of `escapeNumber` is therefore
essentially a no-op, and we should just return the number directly
once we have verified that it is numeric. This reduces the cost of
`escapeNumber` by a factor of ~3 or 4.