universe_create_ports.php: simplify javascript
Use the `onInput` event instead of `onFocus` event to recalculate
the totals from user input. This is more precisely what we want
to do, and simplifies the javascript (because previously we had
to set an event to turn on recalculating at regular intervals,
and then another event to turn it off).

We also remove the javascript function that was used for mines
(i.e. asteroids), for which the calling code was removed in
844dab95be1.