Refactor shop_ship.php
Split shop_ship.php into a part that queries the ship list and
a part that processes the list (shop_ship.inc).

This is the easiest way to change the ship list for different
game types, because then the query in shop_ship.php can be modified
without duplicating the logic in shop_ship.inc.