Move intrinsic planet data from database to PHP
This change replaces constant planet database tables with constant PHP
arrays. Drop the following tables:

* planet_can_build (list of structures that each planet type can build)
* planet_construction (base structure costs)
* planet_cost_good (list of goods for each structure by planet type)
* planet_type (metadata for each planet type)
* planet_type_has_section (menu options for each planet type)

We remove the following functions:

* Globals::getPlanetBuildings
* SmrPlanet::initMaxBuildings
* SmrPlanet::initTypeInfo
* SmrPlanet::initMenuOptions

We add the following classes:

* SmrPlanetType
* SmrPlanetStructureType

We had discussed this in #88, and at the time it was requested to leave
the planet tables alone in case we ever wanted to make an Admin Tool
that would allow admins to modify planets (like in Edit Locations).
However, there is no accountability in a tool like that, and it is
difficult to tell if the database is the same as what you'd expect
in a fresh dev environment (we would have to add another table to
track local changes).

It is much easier to develop with the data stored in class constants,
changes are more transparent, and it's also more performant (OPcache
vs. query overhead).