Add new planet type and planet structure
Planet type: Protoplanet

This planet type supports only less advanced structure types,
since it is designed for Hunter Wars, where the technology of
the Raider-class ships is not available. It cannot build
Generators or Turrets.

Planet structure: Weapon Mount

This structure is designed with the Hunter Wars game in mind,
where the max port level is 6, and so Computers are not available.
The Weapon Mounts take lower level goods to build, but must be
retrofitted with ship weapons.