universe_create_ports.php: fix port total not updating
The javascript responsible for updating the total number of ports
(by level) in realtime had a hard-coded maximum level for the ports.
This stops working if the actual max level is not what is hardcoded.
To fix this, we simply pass the max port level constant to the
javascript function.