Add new planetary structure: Radar Station
The Radar Station is unique in that it requires a ship's scanner
to be built. It increases the accuracy of mounted weapons on the
planet (but can perhaps be transitioned to Terran planets in the
future as well, to help make the planet weapon accuracy function
more consistent between Turrets and Mounted Weapons).