PlanetList.inc: modify to support new structures
With the addition of Weapon Mounts and Radar Stations, the Planet
List page needed some modification to fit all the relevant
information.

We remove separate columns for structures (Gens, Hangars, etc.)
and hardware (shields, armour, etc.) and replace them with
aggregate columns that mimick the Supplies column.

The new columns are "Structures" and "Hardware", which display
the tooltip icon for the item as well as the "current / max"
values for that item.

Add the following convenience methods:
* SmrPlanet::hasStructureType - (bool) does this planet support
                                the given structure type
* SmrPlanetStructureType::getTypes - (list) all possible structure
                                     type ID's