SmrPlanet: change how we choose structure to destroy
Previously, there was a flat chance to destroy each type of structure.

Generators/Bunkers = 50%
Hangars = 33%
Turrets = 17%

With the addition of new structures, we need a more formulaic method
to choose a structure. So we use the following formula:

% Chance = (Num Buildings / Max Buildings) / Build Time

Therefore, you're more likely to lose structure types that you have
a lot of (compared to how many the planet can support), and types
that take longer to build have a lower chance to be lost.

For a max level Terran planet, the chances work out to be similar
to the original constant values:

Generators = 60%
Hangars = 30%
Turrets = 10%

(Added a `getWeightedRandom` function to help pick the structure
to lose given an array of keys and weights.)