Minor ship changes starting Game 122
Thief:
    +10 holds (150 -> 160)

Assassin:
    +1 hp (4 -> 5)

Death Cruiser:
    +1 hp (5 -> 6)
    -25 mines (125 -> 100)

Advanced Carrier:
    -50 shields (750 -> 700)
    +50 armour (100 -> 150)
    -25 drones (275 -> 250)

The changes to the Assassin and ITAC are to help fix the relative
win % chances of the hunter ships. This addresses the most over-
and under-powered hunters.

The Thief and DC are modified in relation to the Assassin. The
Thief gets a very minor boost to trade potential, and the DC trades
a little bit of utility for a more reasonable number of weapons.