preferences.php: move Chat Sharing to game options
Opening the Chat Sharing preferences without first joining a game
was causing a database error, since the Chat Sharing engine code
depends on `$player`. The easiest way to fix this is to add the
link to the game-specific preferences at the top of the page, rather
than with the account-specific preferences below.

This moves the Chat Sharing away from its natural place next to
the Discord ID/IRC Username options, but it is a small price to pay.