validate_processing.php: don't silently skip check
If you clicked "Validate me now!" without entering a validation
code, we would silently skip the validation check (and the account
would still not be validated). This was misleading, so now we
always do the validation check unless "I'll validate later." is
clicked.

We also display any error messages on the validation page itself,
since otherwise you can't get back to it after an error (e.g. an
incorrect validation code is entered) without relogging, which was
not very intuitive.