PlanetList.inc: fix non-shrinking table columns
At least one column of a table has to _not_ have the `shrink`
attribute for any of the `shrink` attributes to have an effect.

We want the planet image (column 1) to shrink, so we remove the
attribute from the "Name", "Owner", and "Build" columns. This
results in a spacing that seems okay.