VoteSite: add PBBG link
The PBBG site can be used to submit reviews for the game, but it
does not have a voting callback, so it cannot be used to give free
turns.

As a result, we do not give a `img_star` or `url_func` for this site,
and we add a `VoteSite::givesFreeTurns` method that we use in the
other methods to avoid calling functions that don't make sense for
non-voting sites.