Generate random validation codes
When creating an account or updating an e-mail address, set the
validation code from a random string instead of a substring of
the session ID (which the user has access to via their cookies
even if the specified e-mail address is invalid).

This prevents a possible exploit for creating accounts with
invalid e-mail addresses.