preferences_processing.php: fix change password bug
Fix a bug that let you change your password even if you entered the
wrong current password.

The SmrAccount::checkPassword method was wrong because it didn't
hash the input password before checking it against the stored hash,
so it always returned "false".

However, the "Current Password" input in the "Change Password"
option in preferences was also wrong in a way that cancelled out
the SmrAccount error, because it would only trigger the "Your current
password is wrong" error if `checkPassword` returned "true"!

To fix this:
* flip the logic in the `checkPassword` call in preferences
* compare the hashed input password to the stored password