Switch to secure encryption for password storing
The md5 hashing algorithm has severe vulnerabilities that do not
make it safe for storing passwords. PHP has a special API for
performing strong password hashing.

We now store the passwords using the default encryption algorithm
(currently bcrypt), and rehash when necessary if any of the defaults
change. The algo and salt are stored as part of the hash with the
encrypted password, which is why we can use `password_verify` to
determine if the input password is correct, even if PHP has changed
the output of a default call to `password_hash`.

We retain the ability to detect md5 hashed passwords, but they will
be automatically upgraded to the new encryption after login.

Modify `account.password` column from `char(32)` to `varchar(255)`
according to the PHP recommendations for storing the output of
`password_hash`. This future-proofs the hash, which could change
length if and when the PASSWORD_DEFAULT constant changes.