Use php-cli-alpine for api-docs image
This reduces the intermediate image size from ~1.2GB (with the
standard php-cli image) to ~0.7GB. All auxilliary php services
now use the php-cli-alpine image.

To get alpine to work with phpDox, we needed to use an updated
iconv extension. This is a bit of a hack, but gets the job done.
See https://github.com/docker-library/php/issues/240.