Choose number of galaxies after game creation
Move the input to choose the number of galaxies out of the game
creation page and into the galaxy creation page. This is useful
for two reasons:

1. It allows us to refactor the game creation table into both a
   creation and editing page (since it will be non-trivial to
   change the number of galaxies).

2. We want the option to create galaxies/sector from an SMR file
   in the galaxy creation page, and it doesn't make sense to
   specify the number of galaxies before that.

Additional HTML5 enhancements:
* Number of galaxies restricted to between 1 and 30.
* Galaxy height/width restricted to between 1 and 100.

Also fixed error resulting from choosing a number of galaxies
bigger than the length of the default galaxy name array. We just
use the empty string for any additional galaxy names.