universe_create_sectors.php: forward to galaxy creation
If you try to edit a game that doesn't have any galaxies, instead
of triggering a database error, simply forward to the galaxy creation
page instead.

This will allow us to continue where we left off if we created a
game but then had an error or timed out while creating galaxies.